UPDATE: use long for ids

+++ src/main/java/mops/businesslogic/DirectoryService.java
@@ -13,7 +13,7 @@ public interface DirectoryService {
      * @param dirId    the id of the folder where the file will be uploaded
      * @param fileInfo the file object
      */
-    void uploadFile(Account account, int dirId, FileInfo fileInfo);
+    void uploadFile(Account account, long dirId, FileInfo fileInfo);
 
     /**
      * Returns all folders of the parent folder.
@@ -22,7 +22,7 @@ public interface DirectoryService {
      * @param dirId   id of the folder
      * @return list of folders
      */
-    List<Directory> getSubFolders(Account account, int dirId);
+    List<Directory> getSubFolders(Account account, long dirId);
 
     /**
      * Creates a new folder inside a folder.
@@ -32,7 +32,7 @@ public interface DirectoryService {
      * @param dirName name of the new folder
      * @return id of the new folder
      */
-    int createFolder(Account account, int dirId, String dirName);
+    int createFolder(Account account, long dirId, String dirName);
 
     /**
      * Deletes a folder.
@@ -41,7 +41,7 @@ public interface DirectoryService {
      * @param dirId   id of the folder to be deleted
      * @return the parent id of the deleted folder
      */
-    int deleteFolder(Account account, int dirId);
+    int deleteFolder(Account account, long dirId);
 
     /**
      * Searches a folder for files.
@@ -51,5 +51,5 @@ public interface DirectoryService {
      * @param query   wrapper object of the query parameter
      * @return list of files
      */
-    List<FileInfo> searchFolder(Account account, int dirId, FileQuery query);
+    List<FileInfo> searchFolder(Account account, long dirId, FileQuery query);
 }
